pyproject.toml
[project]
name = "elevenlabsapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "elevenlabs>=2.5.0",
    "fastmcp>=2.10.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.30.1",
]
